requirements.txt
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
transformers
tqdm
regex
pyyaml
pandas
pydantic
click
timm
einops
gradio==6.9.0
imageio[ffmpeg]
opencv-python
huggingface_hub
facenet-pytorch
numpy
Pillow
ftfy